api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pyjwt==2.8.0
redis==5.0.1
prometheus-client==0.19.0
slowapi==0.1.9
python-multipart==0.0.6
PyYAML==6.0.1
